fix(metrics): prevent negative utilization on late messages (#24073)
* fix(metrics): prevent negative utilization on late messages

This prevents late messages from breaking expected utilization bounds, by clamping all late messages
to the start of reporting period. Messages arrive in order, so that should not break the expected
state (waiting or not waiting). Utilization metric can still be slightly inaccurate on late
messages, but it was never meant to be completely accurate.

Fixes: #24060

* Add changelog entry

* Add more details to changelog

* Update src/utilization.rs

Co-authored-by: Pavlos Rontidis <pavlos.rontidis@gmail.com>

* Use `assert_approx_eq` in utilization tests

---------

Co-authored-by: Pavlos Rontidis <pavlos.rontidis@gmail.com>