chore(transforms): Allow `datadog_search` to use `&LogEvent` directly
The current `datadog_search` matcher interface takes a reference to an `&Event`
and then conditionally searches only logs, returning `false` if the event is not
a log. If we want to search on type `LogEvent`, we need to first wrap it in an
`Event` and then search. By using an `EventRef`, though, we can accomodate
searching on both `&Event` and `&LogEvent` directly, avoiding the extraneous
conversion.