enhancement(security): rework TlsSettings to carry PEM based objects (#23146)
* internal(tlssettings tlsconfig): rework TlsSettings and TlsConfig to use PEM files instead of PKCS12 for FIPS compatibility

TlsSettings used a PKCS12 archive to store the TLS identity information. PKCS12 is not
Federal Information Processing Standard (FIPS) compliant, and there is significant interest
to use vector in such environments.

This change makes PEM based certs & keys the format stored in TlsSettings. For compatibility
when a PKCS12 (DER) archive is presented in the configuration, the code will extract the
components into X509 at load time. This conversion is not FIPS compliant and should be avoided
there. In pratice, most use cases do not use FIPS, so this is perfectly fine to use.

This change was tested on a regular machine, and on a FIPS enabled kernel where we compiled
vector to dynamically link to the system provided OpenSSL. Only the 'fips' and 'base' providers
were configured to be available.

RUSTFLAGS="-C prefer-dynamic" OPENSSL_NO_VENDOR=1 OPENSSL_STATIC=0 \
    cargo build -j8 --release --target x86_64-unknown-linux-gnu \
    --no-default-features --features target-x86_64-unknown-linux-gnu

Key changes:
* IdentityStore now keeps a name, cert, private key, and an optional list of ca certs
* PKCS12 configuration will convert to IdentityStore on load
* TlsSettings::identity() returns the IdentityStore instead of a PKCS12 archive. This
    avoids the need to convert to PCKS12 and thus avoids the FIPS incompatibility for
    future uses

* cargo fmt

* revive PR

* cargo fmt

* add changelog

---------

Co-authored-by: Pavlos Rontidis <pavlos.rontidis@gmail.com>
Co-authored-by: Thomas <thomas.schneider@datadoghq.com>