chore(observability): Introduce `trait NamedInternalEvent` and derive (#24313)
* chore(observability): Introduce `trait NamedInternalEvent` and derive

This change moves the two optional `fn name` methods from the `InternalEvent` and
`RegisterInternalEvent` traits into a new `trait NamedInternalEvent` and adds a
derive procedural macro to make implementation of that trait trivial.

This borrows heavily from #23978 which implemented a `#[internal_event]`
procedural macro, but takes it in a different direction and completes the
implementation.

* Miscellaneous fixes

* Drop `DefaultName` wrapper as it is no longer needed

* Fix Windows service events